+++ mex/common/transform.py
@@ -1,15 +1,16 @@
 import json
 import re
+from collections.abc import Iterable
 from enum import Enum
 from functools import cache
 from pathlib import PurePath
-from typing import Any, Iterable, cast
+from typing import Any, cast
 from uuid import UUID
 
 from pydantic import AnyUrl, SecretStr
 from pydantic import BaseModel as PydanticModel
 
-from mex.common.types import PathWrapper, Timestamp
+from mex.common.types import PathWrapper, TemporalEntity
 
 
 class MExEncoder(json.JSONEncoder):
@@ -27,7 +28,7 @@ class MExEncoder(json.JSONEncoder):
             return obj.value
         if isinstance(obj, UUID):
             return str(obj)
-        if isinstance(obj, Timestamp):
+        if isinstance(obj, TemporalEntity):
             return str(obj)
         if isinstance(obj, PurePath):
             return obj.as_posix()