+++ mex/common/types/email.py
@@ -1,4 +1,4 @@
-from typing import Any, Type
+from typing import Any
 
 from pydantic import GetJsonSchemaHandler
 from pydantic.json_schema import JsonSchemaValue
@@ -11,7 +11,7 @@ class Email(str):
     """Email address of a person, organization or other entity."""
 
     @classmethod
-    def __get_pydantic_core_schema__(cls, _source: Type[Any]) -> core_schema.CoreSchema:
+    def __get_pydantic_core_schema__(cls, _source: type[Any]) -> core_schema.CoreSchema:
         """Get pydantic core schema."""
         return core_schema.str_schema(pattern=EMAIL_PATTERN)